chore(bloc-todos): 🔧 add TodosOverviewSnackbar component
