+++ examples/react/bloc-todos/package.json
@@ -12,11 +12,11 @@
   "dependencies": {
     "@emotion/react": "11.11.1",
     "@emotion/styled": "11.11.0",
-    "@jacobtipp/bloc": "^3.0.1",
-    "@jacobtipp/bloc-concurrency": "^3.0.2",
-    "@jacobtipp/bloc-devtools": "^1.0.2",
-    "@jacobtipp/hydrated-bloc": "^3.0.2",
-    "@jacobtipp/react-bloc": "^3.2.0",
+    "@jacobtipp/bloc": "^3.0.2",
+    "@jacobtipp/bloc-concurrency": "^3.0.3",
+    "@jacobtipp/bloc-devtools": "^1.0.3",
+    "@jacobtipp/hydrated-bloc": "^3.0.3",
+    "@jacobtipp/react-bloc": "3.2.1",
     "@jacobtipp/state": "^1.0.7",
     "@mui/material": "^5.10.17",
     "material-ui-popup-state": "^5.0.4",
