+++ examples/react/bloc-todos/src/features/todos-overview/view/todos-overview.tsx
@@ -9,8 +9,6 @@ import {
   IconButton,
   Toolbar,
   Typography,
-  Button,
-  Snackbar,
   Container,
 } from '@mui/material';
 import Icon from '@mui/material/Icon';
@@ -19,25 +17,19 @@ import { useNavigate } from 'react-router-dom';
 import { TodosOverviewOptionsButton } from '../components/todos-overview-options-button';
 import { TodosOverviewFilterButton } from '../components/todos-overview-filter-button';
 import { TodosOverviewEmptyText } from '../components/todos-overview-empty-text';
-import { useMemo, useState } from 'react';
 import { Todo } from '../../../packages/todos-client/model/todo';
-import {
-  BlocProvider,
-  useBloc,
-  useBlocListener,
-  useRepository,
-} from '@jacobtipp/react-bloc';
+import { BlocProvider, useBloc, useRepository } from '@jacobtipp/react-bloc';
 import { createSelector } from 'reselect';
 import { TodosOverviewBloc } from '../bloc/todos-overview.bloc';
 import {
   TodosOverviewSubscriptionRequested,
-  TodosOverviewUndoDeletionRequested,
   TodosOverviewTodoDeleted,
   TodosOverviewTodoCompletionToggled,
 } from '../bloc/todos-overview.event';
 import { TodosOverviewState } from '../bloc/todos-overview.state';
 import { TodosOverviewFilter } from '../model/todos-overview-filter';
 import { TodosRepository } from '../../../packages/todos-repository/todos-repository';
+import { TodosOverViewSnackbar } from '../components/todos-overview-snackbar';
 
 const todoFilterMap = (todo: Todo, filter: TodosOverviewFilter) =>
   filter === 'all'
@@ -69,61 +61,18 @@ export default function TodosOverviewPage() {
         )
       }
     >
+      <TodosOverViewSnackbar />
       <TodosOverviewView />
     </BlocProvider>
   );
 }
 
 export function TodosOverviewView() {
-  const [isSnackbarOpen, openSnackbar] = useState(false);
-  const [snackbarMessage, setSnackbarMessage] = useState('');
   const [filteredTodos, { add }] = useBloc(TodosOverviewBloc, {
     selector: filteredTodosSelector,
   });
 
   const navigate = useNavigate();
-
-  const handleCloseSnackbar = (
-    _event: React.SyntheticEvent | Event,
-    reason?: string
-  ) => {
-    if (reason === 'clickaway') {
-      return;
-    }
-
-    openSnackbar(false);
-  };
-
-  const action = useMemo(() => {
-    return (
-      <Button
-        color="secondary"
-        size="small"
-        onClick={(e) => {
-          handleCloseSnackbar(e);
-          add(new TodosOverviewUndoDeletionRequested());
-        }}
-      >
-        UNDO
-      </Button>
-    );
-  }, [add]);
-
-  useBlocListener(TodosOverviewBloc, {
-    listenWhen(previous, current) {
-      return (
-        previous.data.lastDeletedTodo !== current.data.lastDeletedTodo &&
-        current.data.lastDeletedTodo !== undefined
-      );
-    },
-    listener(_bloc, state) {
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      const deletedTodo = state.data.lastDeletedTodo!;
-      setSnackbarMessage(`Todo "${deletedTodo.title}" deleted.`);
-      openSnackbar(false); // close snackbar if already open
-      openSnackbar(true); // open snackbar
-    },
-  });
 
   return (
     <>
@@ -191,14 +140,6 @@ export function TodosOverviewView() {
           <TodosOverviewEmptyText />
         )}
       </Container>
-      <Snackbar
-        open={isSnackbarOpen}
-        autoHideDuration={6000}
-        onClose={handleCloseSnackbar}
-        message={snackbarMessage}
-        action={action}
-        sx={{ bottom: { xs: 100, sm: 100 } }}
-      />
     </>
   );
 }
